before patching

+++ src/main/java/ru/lukin/spring/nic_zadacha/DTO/SubContractDTO.java
@@ -3,6 +3,7 @@ package ru.lukin.spring.nic_zadacha.DTO;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Getter
@@ -11,10 +12,10 @@ public class SubContractDTO {
     private Long id;
     private String name;
     private String contractType;
-    private Integer amount;
+    private BigDecimal amount;
     private LocalDate plannedStartDate;
     private LocalDate plannedEndDate;
     private LocalDate actualStartDate;
     private LocalDate actualEndDate;
-    private Long counterpartyId; // Только id контрагента
+    private Long counterpartyId;
 }
