+++ src/main/java/ru/lukin/spring/nic_zadacha/model/ContractStage.java
@@ -1,21 +1,23 @@
 package ru.lukin.spring.nic_zadacha.model;
 
-import com.fasterxml.jackson.annotation.JsonIdentityInfo;
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import com.fasterxml.jackson.annotation.ObjectIdGenerators;
-import jakarta.persistence.*;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+import jakarta.persistence.GenerationType;
 import lombok.Getter;
 import lombok.Setter;
 
 import java.math.BigDecimal;
 import java.time.LocalDate;
-import java.util.List;
 
 @Entity
 @Getter
 @Setter
 @Table(name = "contract_stage")
 public class ContractStage {
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
@@ -25,9 +27,9 @@ public class ContractStage {
     private LocalDate plannedEndDate;
     private LocalDate actualStartDate;
     private LocalDate actualEndDate;
-    private Integer amount;
-    private Integer materialCostsPlan;
-    private Integer materialCostsActual;
-    private Integer salaryCostsPlan;
-    private Integer salaryCostsActual;
+    private BigDecimal amount;
+    private BigDecimal materialCostsPlan;
+    private BigDecimal materialCostsActual;
+    private BigDecimal salaryCostsPlan;
+    private BigDecimal salaryCostsActual;
 }
