+++ src/main/java/ru/lukin/spring/nic_zadacha/model/SubContract.java
@@ -1,9 +1,17 @@
 package ru.lukin.spring.nic_zadacha.model;
 
-import jakarta.persistence.*;
+
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Entity
@@ -11,12 +19,13 @@ import java.time.LocalDate;
 @Setter
 @Table(name = "sub_contract")
 public class SubContract {
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
     private String name;
     private String contractType;
-    private Integer amount;
+    private BigDecimal amount;
     private LocalDate plannedStartDate;
     private LocalDate plannedEndDate;
     private LocalDate actualStartDate;
